#3 fix: 레이아웃 설정 홈 뷰로 이동
